move executeErrorWorkflow out of workflow-execute-additional-data
